+++ src/main/java/com/whalespottingjava/controllers/AdminController.java
@@ -1,5 +1,7 @@
 package com.whalespottingjava.controllers;
 
+import com.whalespottingjava.models.database.Sighting;
+import com.whalespottingjava.models.requests.AdminApprovalRequest;
 import com.whalespottingjava.services.SightingService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -8,7 +10,10 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
+import java.util.List;
+import java.util.Optional;
 
 @Controller
 @PreAuthorize("hasRole('ADMIN')")
@@ -30,8 +35,17 @@ public class AdminController {
 
     @PostMapping("/admin")
     @ResponseStatus(HttpStatus.CREATED)
-    public String postSightings(Model model) {
-        
+    public String postSightings(@RequestBody List<AdminApprovalRequest> list, Model model) {
+        for (AdminApprovalRequest sighting: list) {
+            if (sighting.getApproved() == false) {
+                sightingService.deleteRejectedPendingSighting(sighting.getId());
+            }
+            if (sighting.getApproved() == true) {
+                Optional<Sighting> currentSighting = sightingService.getSightingById(sighting.getId());
+                sightingService.addSighting(currentSighting);
+            }
+            
+        }
         return "admin";
     }
 }
